move audio toolkit into src-tauri to fix permissions
