+++ .gitignore
@@ -23,7 +23,5 @@ dist-ssr
 *.sln
 *.sw?
 
-src-tauri/resources/models
-
 /target/
 recording_*
