+++ src-tauri/Cargo.toml
@@ -5,6 +5,7 @@ description = "Handy"
 authors = ["cjpais"]
 edition = "2021"
 license = "MIT"
+default-run = "handy"
 
 [profile.dev]
 incremental = true # Compile your binary in smaller steps.
@@ -18,6 +19,10 @@ incremental = true # Compile your binary in smaller steps.
 name = "handy_app_lib"
 crate-type = ["staticlib", "cdylib", "rlib"]
 
+# [[bin]]
+# name = "cli"
+# path = "src/audio_toolkit/bin/cli.rs"
+
 [build-dependencies]
 tauri-build = { version = "2", features = [] }
 
@@ -30,10 +35,11 @@ serde_json = "1"
 tauri-plugin-clipboard-manager = "2"
 tauri-plugin-macos-permissions = "2.0.4"
 rdev = { git = "https://github.com/rustdesk-org/rdev" }
-cpal = "0.15.3"
+cpal = "0.16.0"
 whisper-rs = { version = "0.13.2", features = ["whisper-cpp-log"] }
 anyhow = "1.0.95"
-rubato = "0.16.1"
+rubato = "0.16.2"
+hound = "3.5.1"
 env_logger = "0.11.6"
 log = "0.4.25"
 tokio = "1.43.0"
@@ -45,7 +51,6 @@ tauri-plugin-process = "2"
 rodio = "0.20.1"
 reqwest = { version = "0.11", features = ["json", "stream"] }
 futures-util = "0.3"
-audio-toolkit = { path = "../crates/audio-toolkit" }
 
 [dependencies.ort-sys]
 version = "=2.0.0-rc.9"
